Extract shareable code to a service
